*** These changes are experimental and not ready for stringent code *** *** review.  I am committing specifically to facilitate discussions *** *** that Max and Erik and I are having on irc.                      ***
These changes represent no change in functionality.  My goal is to
stop passing around a gazillion variables and using a bunch of little
helper functions.  Instead, I present to you the CVSRevision class,
which represents--yep, a CVS Revision.  It will not only absorb the
bits of data that are floating through cvs2svn, but also the behaviors
relating to cvs revisions.

This is the first step to using CVSRevisions throught cvs2svn.py.
Much more to come.

* cvs2svn.py: Gutting and cleaning.