Commit up Fitz's changes to commit group aggregation.
Now hear this, now hear this: 

We know this makes the test suite start failing on the branch.  This
is very much a work in progress.  A lot of this stuff will be changed
very soon.  I'm committing it because Fitz had to go, and I wanted to
continue where he left off.  He handed me the patch as he walked out,
then I tweaked it a bit, wrote up a log message, and committed.  So
expect more churn in the very near future.

* cvs2svn.py
  (CollectData.tree_completed): Comment out a resync print.
  (CVSRevisionDatabase): Add a comment questioning the necessity of
    this database.
  (RepositoryHead): New singleton class.
  (RevisionMappingDatabase): New class.
  (CVSCommit): New class.
  (SVNCommit): New class, still skeletal.
  (SVNRevNum): New singleton class.
  (CVSRevisionAggregator): New class.
  (pass8): Instantiate a CVSRevisionAggregator, use it to process c_rev.