Issue 44 - branch filling algorithm could be smarter.
Rewrite the symbolic name filler.  The new code loops over the
destination tree and considers all sources for each destination path.
By sorting the sources in descending score order we can make the copy
from the source that has the highest score.  The score is not the
optimal metric to select copy source, but it's reasonably good.

The new code makes better copies, is slightly faster, and it fixes a
bug in the conversion of a repository that I unfortunately cannot
share.  The bug is not present on trunk and the symptom is missing
entries in tags and branches.

* may-04-redesign/cvs2svn.py:
  (_path_join): New helper function.

  (SymbolicNameFillingGuide.get_best_revnum,
   SymbolicNameFillingGuide._best_rev): Return the best score as well
  as the revnum.

  (FillSource): New class to store a fill source for the symbolic
  name filler.

  (SVNRepositoryMirror:._fill_symbolic_name): Update to match the
  new _fill function.

  (SVNRepositoryMirror:._fill): Rewritten.

  (SVNRepositoryMirror:._list): New function to return the items
  in a directory.

* may-04-redesign/run-tests.py:
  (compose_tag_three_sources): Update test to match new filler code.