Change Log.check_changes() to accept a list of tuples instead of a dict.
Using uninterpolated strings as the keys for a dictionary is not
really correct.  And the (debatable) efficiency benefit of passing
CHANGED_PATHS as a dict has been lost by now anyway, because string
interpolation has to be done on the keys before it can be used.
Therefore, check_changes() has been rewritten to accept a list of
(path,op) tuples.

* run-tests.py (Log.check_changes, Log.changes): Accept a list of
  tuples instead of a dictionary for the CHANGED_PATHS argument.
  Callers changed accordingly.