Simplified algorithm of SymbolicNameFillingGuide....score_revisions().
The new algorithm is sensible, it implements the contract documented
in the old docstring, and it produces the exact same results as the
old algorithm for all cases that arise in the test suite.  But
honestly, I didn't make the effort to understand the impenetrable old
code, so I can't be 100% sure that the results will be identical in
all cases.

* cvs2svn (SymbolicNameFillingGuide.get_best_revnum.score_revisions):
  Simplified algorithm.