Changed the line-oriented data files to be null-terminated.
This was a well-intentioned but silly attempt to make cvs2svn support
filenames containing control characters.  The main problem was that
filenames with newlines couldn't be handled within the
cvs2svn-data.revs data file and its relatives, because those files
were line-based (so the newline within the filename caused a premature
end-of-record).

This work was pointless because it turns out that Subversion itself
cannot handle control characters in filenames.

Nevertheless I am checking this work into a branch (which I will soon
delete) to keep a record of what would have to be changed to make
cvs2svn support crazy CVS filenames.

This changed version was lightly tested.  All old test cases pass, and
the test case that I added to test this new feature gets to pass8
before failing due to svnadmin's limitations.

* cvs2svn (NullTerminatedRecordReader): Read a file one
  null-terminated record at a time.

  (sort_file): Pass the '-z' option to sort.

  (FileDataCollector.set_revision_info, SymbolingsLogger._log, pass2):
  Write records with terminating '\0' instead of '\n'.

  (CollectData.__init__): Open self.revs in binary mode.

  (SymbolingsLogger.__init__): Open self.symbolings in binary mode.

  (SymbolingsReader.__init__,
  SymbolingsReader.filling_guide_for_symbolpass2, pass4, pass5,
  pass7.generate_offsets_for_symbolings): Read files using
  NullTerminatedRecordReader.

* run-tests.py (ctrl_char_in_filename): New test for filenames
  containing control characters.

* design-notes.txt: Noted that files are null-terminated.  Updated
  text to refer to "records" instead of "lines".