Turn TagsDatabase into a set-like class.
Previously, TagsDatabase was a factory function that returned an
SDatabase instance.  This was broken, because client code tried to set
items in this database to None (whereas an SDatabase generically only
allows strings).

Instead of changing client code to store '' to the database, I changed
TagsDatabase into a set-like class so that clients don't have to worry
about what values to set.  The class now implements a tiny subset of
the "set" interface defined in Python 2.4.

This change seems to help Python 2.2 compatibility, perhaps because
gdbm allows None values in later Python versions (?)

* cvs2svn (TagsDatabase): Rewritten as described above.

  (PersistenceManager.get_svn_commit,
  SVNRepositoryMirror._fill_symbolic_name, pass2): Use TagsDatabase's
  new interface.