Classes for passes.
Added a new base class, Pass, representing a conversion pass.  Derived
Pass1 through Pass8 from it.  The classs have a run() method which
contains the code that used to be in the pass* functions.

For now, the class object is used as the key to control cleanup, but
this will change shortly.

* cvs2svn (Pass, Pass1, Pass2, Pass3, Pass4, Pass5, Pass6, Pass7,
  Pass8): New classes, with code moved from pass1(), pass2(), pass3(),
  pass4(), pass5(), pass6(), pass7(), and pass8().  All users of these
  classes changed accordingly.

  (convert): Call pass.run() instead of invoking the pass as a
  function.  Use pass.__class__ as the key passed to
  cleanup.cleanup().