Get rid of CVSRevision.prev_timestamp and next_timestamp.
When they are needed, fetch them from the database.  This eliminates
the need for TWEAKED_TIMESTAMP_DATABASE, which was only there to bring
prev_timestamp and next_timestamp up to date with the values stored in
the actual previous and next CVSRevision records.

I measured the time impact on my "standard database" and it was
negligible.

* cvs2svn_lib/passes.py (ResyncRevsPass.register_artifacts): Eliminate
  TWEAKED_TIMESTAMPS_DB.

  (ResyncRevsPass._fix_prev_next_timestamps): Remove method.

  (ResyncRevsPass.run): Eliminate TWEAKED_TIMESTAMPS_DB.  Access and
  modify timestamps directly in the effected CVSRevision records.

* cvs2svn_lib/config.py (TWEAKED_TIMESTAMPS_DB): Removed constant.

* cvs2svn_lib/cvs_revision.py (CVSRevision.prev_timestamp,
  CVSRevision.next_timestamp): Removed fields.

* cvs2svn_lib/collect_data.py (FileDataCollector.set_revision_info):
  Don't need prev_timestamp and next_timestamp to construct a
  CVSRevision.