StrictSymbolStrategy works via a list of rules.
Rules are tried in turn until one matches the symbol.  Note that this
intentionally changes the program behavior in the following way: rules
are applied in the order that they are specified on the command line
(via --exclude, --force-branch, or --force-tag options); the first
rule that matches a symbol is the one that is applied.

* cvs2svn_lib/symbol_strategy.py (match_regexp_list): Function
  removed.

  (ExcludeSymbol, StrategyRule, RegexpStrategyRule): New classes.

  (StrictSymbolStrategy): Updated docstring.

  (StrictSymbolStrategy.excludes,
  StrictSymbolStrategy.forced_branches,
  StrictSymbolStrategy.forced_tags): Remove members.

  (StrictSymbolStrategy._rules): New member.

  (StrictSymbolStrategy._compile_re): New helper method

  (StrictSymbolStrategy.add_exclude,
  StrictSymbolStrategy.add_forced_branch,
  StrictSymbolStrategy.add_forced_tag): Simply add a rule to
  self._rules.

  (StrictSymbolStrategy._get_symbol): New method.

  (StrictSymbolStrategy.get_symbols): Determine how to convert symbol
  via self._rules.