Fix a nasty little inconsistency between comment and practice.
According to the docstring for SVNRepositoryMirror.delete_path(), the
SVN_PATH must exist.  Indeed, all calls to that method from within the
same source file seem to satisfy the requirement.  However, the
requirement was not enforced.  If a non-existent path was passed to
the old version of delete_path(), it created a writable node for the
parent directory but stopped there (without an error).  Thus the old
version didn't fail for nonexistent directories, but it didn't really
handle them correctly, either.

In fact, there were cases in the test suite when delete_path() was
called by SVNPrimaryCommit.commit() with a path that did not exist
(namely in a test that has a file with two OP_DELETEs in a row in the
same branch).  Therefore, I protected this call to delete_path() with
"if repos.path_exists(...)".

SVNPostCommit.commit() also calls delete_path() in a way whose
correctness is not obvious.  So I also added the same check there.

With these changes it is clear that delete_path() is only called with
existing paths.  This makes it possible to change _fast_delete_path()
to also require existing paths.  I made this change.

Unfortunately, the path_exists() check is expensive; it can involve
multiple database lookups.  So the current situation is not at all
attractive, and we should try to improve the situation.

* cvs2svn_lib/svn_commit.py (SVNPrimaryCommit.commit,
  SVNPostCommit.commit): Check that the path exists before calling
  repos.delete_path().

* cvs2svn_lib/svn_repository_mirror.py
  (SVNRepositoryMirror._fast_delete_path): Add requirement that the
  path to be deleted exists, and remove the do-nothing option if it
  doesn't.