Only break SymbolChangeset loops where the OrderedChangesets are retrograde.
The OrderedChangeset's order is already carved in stone, so it is not
helpful to break a SymbolChangeset that connects two OrderedChangesets
that are already in order.  Instead, we now break any segments of
SymbolChangesets that connect two OrderedChangesets in the wrong
order.

This fixes 8 of the 12 test suite failures that resulted from the
change to using symbol dependencies.

* cvs2svn_lib/passes.py
  (BreakCVSSymbolChangesetLoopsPass.break_segment): New method.

  (BreakCVSSymbolChangesetLoopsPass.break_cycle): Break the segments
  of SymbolChangesets that are bracketed by retrograde
  OrderedChangesets (if there are any).