Add a new test case.
The test tickles a problem that I was having with an uncommitted
version of better-symbol-selection when BRANCH2 is grafted onto
BRANCH1.  The problem is that the copy source directory doesn't exist
at the time that cvs2svn tries to fill BRANCH2.

The test was inspired by a problem in my usual gcc test repository.

* run-tests.py (branch_from_empty_dir): New test.

  (test_list): Add new test to list.

* test-data/branch-from-empty-dir-cvsrepos: Repository to support new
  test.