Look for simple retrograde BranchChangesets and break them intelligently.
Check which BranchChangesets are trivially retrograde--that is, those
that directly connect OrderedChangesets in the wrong order, not
considering indirect connections through other BranchChangesets.
Split each one up in an intelligent way.

This should hopefully cover the large majority of cycles closed by
BranchChangesets.  (The rest are handled by a generic cycle-breaking
algorithm.)

* cvs2svn_lib/passes.py
  (BreakAllChangesetCyclesPass._split_retrograde_changeset,
  BreakAllChangesetCyclesPass._split_if_retrograde): New methods.

  (BreakAllChangesetCyclesPass.ordinals): New member.

  (BreakAllChangesetCyclesPass.run): Fill new ordinals member.  Loop
  through all branch changesets looking for and breaking those that
  are simply retrograde.