Rewrite the bookkeeping code for --use-internal-co.
Keep track of the bookkeeping information in a series of TextRecord
classes.  The module now reads almost all of the information that it
needs out of CVSFileItems instances.  Made the reference counting more
explicit and straightforward.

This will make it easier to modify a file's dependency topology
without having to modify checkout_internal each time.

* cvs2svn_lib/checkout_internal.py: Rewrote much of the module.
  Please see the file comments for details.

* run-tests.py (test_list): Remove the XFail annotation from
  leftover_revs test, since it passes again.