Remove the special handling of an empty SymbolFillingGuide.
Previously, if the first action on a branch was to add a file, there
was a symbol fill with an empty SymbolFillingGuide, which resulted in
the branch directory being created via a copy from trunk using
SVNRepositoryMirror._fill_empty_branch().

But due to recent changes, if a file is added on a branch, the
corresponding CVSBranch is deleted, severing the branch from trunk,
and causing the file to be added on the branch without first requiring
a symbol fill.  Therefore, SVNRepositoryMirror._fill_empty_branch()
was never used anymore.

The net result of these accumulated changes is that if the very first
action on a branch is to add a file, then the branch is brought into
existence using the equivalent of "svn mkdir" rather than "svn mv".
This arguably omits information about the source of the branch.  But I
think it is OK for the following reasons:

1. This case should be rare, because typically a branch is created by
   tagging existing files rather than by creating a new file on the
   branch.

2. Copying the branch from trunk is not really correct anyway.  It
   could be that the branch was created from another branch, in which
   case it would make more sense to bring the branch into existence by
   copying it from the source branch.  But the source branch cannot be
   determined from the single, added file.  The only hope would be to
   copy from the branch's preferred parent (which is determined
   globally).  But the preferred parent has not necessarily been
   created by the time of the file-add commit, so even more special
   cases would be needed.

* cvs2svn_lib/svn_repository_mirror.py
  (SVNRepositoryMirror.fill_symbol): Treat the case of an empty
  SymbolFillingGuide as a fatal InternalError.

  (SVNRepositoryMirror._fill_empty_branch): Remove method.