Never treat a delete revision as the original source of another revision.
It is not possible to retrieve the contents of a deleted revision
using CVS, so don't allow a deleted revision to be used as the
"original source" of another revision.  Also, don't write blobs for
deleted revisions.

This fixes a bug in cvs2git output when using non-inline blobs,
wherein a revision that followed a deleted revision usually had no
contents.

Suggested by: Erik Änggård <erik.anggard@gmail.com>

* cvs2svn_lib/git_revision_recorder.py
  (GitRevisionRecorder._get_original_source): Don't consider
  CVSRevisionDeletes to be possible sources.  Update and improve
  docstring.

  (GitRevisionRecorder.record_fulltext): Don't record blobs for
  CVSRevisionDeletes.  Update and improve docstring.