Handle times in the local timezone because Windows doesn't respect TZ.
Patch by: James Abbatiello <abbeyj@gmail.com>
(with comment revisions by me)

James's explanation:

The run-tests.py tries to set things up so that svn will output
timestamps in UTC instead of using the local timezone.  The comment
says "I have no idea if this works on Windows".  Well it doesn't work
on Windows unfortunately.  Timestamps keep getting output in the local
timezone.  There doesn't seem to be any way to override this behavior
on Windows.  The timestamp_chaos test ends up failing with all the
times off by a constant number of hours.  It doesn't seem too hard to
handle timestamps in the local timezone and in fact the comments
already make reference to doing this.  They don't seem to correctly
describe what is going on now with svn configured to output UTC times.
A small change to interpret the times that are hardcoded in
run-tests.py as UTC instead of local allows the test to run with svn
outputting local timestamps.  As a bonus this makes the existing
comments correct again.

* run-tests.py (timestamp_chaos): Use calendar.timegm() to convert
  times properly from UTC strings into seconds since the epoch.

  ("__main__"): Do not set the TZ environment variable, since it does
  not affect the Subversion client under Windows.