Adjust test suite to work with the current version of svntest.
With this change, the test suite works with all versions of
Subversion, including Subversion >= 1.6.0 (which was broken due to a
change in the output format of "svn proplist -v").  Thus issue #124 is
fixed.

This change relies on r38486 in the Subversion svntest module.  It
changes the svn:external to pull in the HEAD version of svntest; we
can try to track svntest again.

* run-tests.py (Cvs2SvnTestFunction): New class, derived from
  TestCase.  It can be used as a decorator for no-argument test
  functions.

  (Cvs2SvnTestCase.__init__): Rename the "description" parameter to
  "doc".  Pass it to the TestCase constructor.

  (Cvs2SvnTestCase.get_description): Remove method.

  (Cvs2SvnTestCase.get_sandbox_name): New method.

  (Cvs2SvnPropertiesTestCase.run): Accept a sandbox parameter.

  (ignore): Fix the end-of-line characters in the expected property
  values to work with the new svntest output.

  (<rest of file>): Decorate all test functions with
  Cvs2SvnTestFunction).  Accept a sandbox argument to all other test
  cases' run() methods.

  (svntest): Change svn:external to pull in the HEAD version of
  svntest from the Subversion project.