Add an --include-empty-directories option.
When this option is set, treat an empty directory within the CVS
repository as a part of the source tree, and create it in the SVN
repository at the point when the directory's parent is created.  (An
empty directory that is directly in a project root directory is
created when the project is initialized.)

Show, in cvs2svn-example.options, how to accomplish the equivalent of
--include-empty-directories via an options file.

Add tests of this new feature.

This commit fixes issue #31.