Try using the -Q cvs global option before trying -q.
Thomas Gardiner reported that CVS was emitting a warning to stderr:

> cvs: WARNING: Read-only repository access mode selected via `cvs
> -R'. Using this option to access a repository which some users write
> to may cause intermittent sandbox corruption.

This warning made cvs2svn consider the cvs command to have failed.

Daniel Black reported that changing the global option "-q" to "-Q"
fixed the problem for him.  The "-Q" option seems to have been added
to CVS in 1991, so it should be safe.  But just to be sure, if -Q
doesn't work, -q is still tried.