Support cProfile profiler.
Patch by: Jon Foster <jon.foster@cabot.co.uk>

According to the Python 2.5 docs, the "hotshot" profiler is deprecated
and may be removed in a future release of Python.  Those docs
recommend the "cProfile" profiler, which is new in Python 2.5.

This patch makes cvs2svn use cProfile if available, and fall back to
the existing hotshot support if cProfile isn't available.