A program for generating blobs externally to main cvs2git program.
This program can generate a git data blob file in git-fast-import
format for arbitrary revisions from RCS files.  The revision fulltexts
are generated using the RCSStream class, the same as is used for the
--use-internal-co option; therefore it is very fast.  And since the
revisions are written to the blobfile in the order they are generated,
most of the work can be done in RAM (though some temporary storage is
used for branches), making it faster still.  And since this program
can run in parallel to the main cvs2git program, it allows the
conversion to take advantage of multiple CPUs.

Subsequent commits will allow this program to be used as part of a
cvs2git conversion.