Push the illegal filename character check down from OutputOption.
For Subversion, do that part of the check in SVNOutputOption.

Git accepts any characters, so don't do an analogous check in
GitOutputOption.

Bazaar requires filenames to be valid UTF8, but we can't check that
here yet because lots of filenames will probably fail until we (at
least) try to use the selected encodings to convert to Unicode.  So
more work still needs to be done.

Mercurial's constraints still need to be investigated.