verify-cvs2svn.py: Do not blithely discard the last character from each line.
SvnRepos.list() was discarding the last character of each line of
output from "svn ls $URL", under the assumption that the last character
would be a "/".  But that assumption is incorrect if the directory
contains a file (as opposed to only subdirectories).  So change the
code to use rstrip() to remove trailing slash characters.