generate_blobs.py: Be careful if deleting revision 1.1.
It is possible that *no* revisions will be needed from an RCS file.
For example, if there was never a commit on the main branch (so
revision 1.1 is dead), and all branches revisions are excluded from
the conversion, then no revisions will be needed from the file.

This situation led to a problem in WriteBlobSink.tree_completed(),
namely that when deleting revision 1.1, it tried to adjust the
is_needed() status of its "base" revision.  Since revision 1.1 doesn't
have a base revision, the lookup failed with a KeyError.

So, if the base revision of the revision being deleted is None, then
skip further processing.

Problem reported by: Sam Hanes <sam@maltera.com> (in a wonderfully
thorough bug report).